Feature logo + multi favicons (#10)
* Logo alteration: replaced the letter A with L.

* L was moved 5 px up.

* Added a routerLink to the Listify logo, so now it redirects to the homepage when on click.

Version updated.

* Pushing again to change the logo (the conflict was unresolved, trying to fix it.)

* I added more favicons, so they look good across all platforms and browsers.

* 1. I changed the order of favicons to ascending.
2. I moved all the favicons to /assets/image directory. I chose to create an additional image directory, to keep the images separate from the other assets that will soon appear in the asset directory.
3. As for the absolute/relative paths:
We use absolute paths to point out of the website. For the internal links, we use relative paths. To set up the navigation, we should specify the full domain name as the base href, for example: <base href="https://www.listify.com/". To access the files from the root directory, we use "/" to start the path. For example: href="/assets/images/our-image.jpg".

However, in this case, when I specified the current project domain https://notes.eager-code.eu/, and applied the previously mentioned rules, my localhost became empty, because I cant reach the domain, I guess, without changes being pushed. So what I did, I left the base with a "/", which means "calculated with respect to base URL's host”. What I understand, is that the browser would take any host URL (local host or not), then puts a "/" and the relative path starts from there.
This is why in the relative paths below I did not add another dash, as I would normally do with a full URL in the base.

* 1. Images moved to assets/images
2. .xml and webmanifest moved to assets
3. Changed paths in index.html
4. 16x16 png was removed.

On my localhost, I couldn't see the favicon on the browser, after I did the changes. I created a dist folder, and the favicons and xlm and webmanifest files where there.

I couldn't google out why this happened. It is interesting to know if it works.

* Use absolute path for assets

* Changing Android favicon background to transparent.

* Fixing iphone icons.

Co-authored-by: Rutenis Turcinas <rutenis.turcinas@devbridge.com>